// Frontend/AngularProject/src/app/clients/common/model/conversationDto.ts
/**
 * Common api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ConversationType } from './conversationType';
import { UserListDto } from './userListDto';

export interface ConversationDto { 
    id?: number;
    name?: string;
    type?: ConversationType;
    users?: Array<UserListDto>;
    inserted?: Date;
}