// Frontend/AngularProject/src/app/clients/common/model/messageDto.ts
/**
 * Common api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentDto } from './attachmentDto';

export interface MessageDto { 
    id?: number;
    text?: string;
    userId?: number;
    sent?: Date;
    conversationId?: number;
    attachments?: Array<AttachmentDto>;
}